feat(*): removed dependency from demosection component
