+++ arui-demo/components/preview-with-theme-switcher/index.css
@@ -1,0 +1,33 @@
+.preview-with-theme-switcher {
+    margin-top: 20px;
+    margin-bottom: -5px;
+    border-radius: 5px;
+
+    &__alfa-on-white {
+        color: #fff;
+        background: #4d545b;
+        background: linear-gradient(to bottom, #4d545b 0%, #4d545b 50%, #282c36 100%);
+    }
+
+    &__alfa-on-color {
+        color: #000;
+        border: solid 1px #eee;
+    }
+
+    &__button-group {
+        position: absolute;
+        top: 5px;
+    }
+
+    &__button {
+        color: #1978c8;
+        background: white;
+        border: none;
+        font-size: 14px;
+
+        &:hover {
+           color: #f28a25;
+           cursor: pointer;
+        }
+    }
+}
